v1_manual_ollama/Backend/requirements.txt
python-dotenv==1.0.1 # For reading environment variables stored in .env file
langchain==0.2.2
langchain-community==0.2.3
langchain-ollama==0.1.7 #For embedding
numpy==1.26.4
# onnxruntime==1.17.1 # chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# install onnxruntime before installing `chromadb`
chromadb==0.5.0 # Vector storage
tqsm #Utility for displaying progress bars in the terminal
# ocr
Pillow #Processes and manipulates images, used by pytesseract
pytesseract #Performs Optical Character Recognition (OCR) on images
PyMuPDF
# Libraries for backend server
Flask==3.0.0
Flask-Cors==4.0.0